MySQL workbench database containing songs
This is for the search page. If you want to use it, update the SearchServlet.java url to the database name (if you are using your own database, otherwise keep it), username (usually root), and password to your own for accessing the sql database. Afterwards run the sql file in the workbench (using lightning bolt icon) and the search.jsp should show the songs in the table.